Extracted `mdbook serve` to separate file/module
